Default to thread first. #:as now restarts threading.
Thanks to the change in #:as semantics we can now use any forms as
threaded clauses e.g. if, when, cond, let, etc. Effectively #:as binds
current value of ~ to an id which you can use in the following clause,
but you can't use ~. Value it computes is used to continue threading.